separate out stories for CheckboxGroup and RadioButtonGroup
